+++ dualsimplexclass.h
@@ -7,8 +7,8 @@ class DualSimplexClass : public LPMethod
 {
     Q_OBJECT
 public:
-    explicit DualSimplexClass(const QVector<float>& objFuncCoeffVector, const QVector<QVector<float>>& constrCoeffMatrix,
-                              const QVector<int>& signs, const QVector<float>& plans, QObject *parent = nullptr);
+    explicit DualSimplexClass(const QVector<double>& objFuncCoeffVector, const QVector<QVector<double>>& constrCoeffMatrix,
+                              const QVector<int>& signs, const QVector<double>& plans, QObject *parent = nullptr);
     DualSimplexClass(const LpStructure& otherStructure);
     DualSimplexClass();
