+++ lpmethods.h
@@ -4,25 +4,25 @@
 #include <QObject>
 
 struct LpStructure {
-    QVector<float> objFuncCoeffVector;
-    QVector<QVector<float>> constrCoeffMatrix;
+    QVector<double> objFuncCoeffVector;
+    QVector<QVector<double>> constrCoeffMatrix;
     QVector<int> signs;
-    QVector<float> plans;
+    QVector<double> plans;
     int leadRowIndex;
     int leadColIndex;
     QVector<int> baseIndexes;
-    float resultValue;
-    QVector<float> ratio;
-    QVector<float> lastRow;
-    float leadElement;
+    double resultValue;
+    QVector<double> ratio;
+    QVector<double> lastRow;
+    double leadElement;
 };
 
 class LPMethod : public QObject
 {
     Q_OBJECT
 public:
-    explicit LPMethod(const QVector<float>& objFuncCoeffVector, const QVector<QVector<float>>& constrCoeffMatrix,
-                       const QVector<int>& signs, const QVector<float>& plans, QObject *parent = nullptr);
+    explicit LPMethod(const QVector<double>& objFuncCoeffVector, const QVector<QVector<double>>& constrCoeffMatrix,
+                       const QVector<int>& signs, const QVector<double>& plans, QObject *parent = nullptr);
 
     LPMethod(const LpStructure& otherStructure);
 
@@ -31,7 +31,7 @@ public:
 protected:
     LpStructure structure;
 
-    float leadingElement;
+    double leadingElement;
 
 public:
     virtual bool SolveOneStep();
