+++ simplexclass.h
@@ -8,8 +8,8 @@ class SimplexClass : public LPMethod
 {
     Q_OBJECT
 public:
-    explicit SimplexClass(const QVector<float>& objFuncCoeffVector, const QVector<QVector<float>>& constrCoeffMatrix,
-                             const QVector<int>& signs, const QVector<float>& plans, QObject *parent = nullptr);
+    explicit SimplexClass(const QVector<double>& objFuncCoeffVector, const QVector<QVector<double>>& constrCoeffMatrix,
+                             const QVector<int>& signs, const QVector<double>& plans, QObject *parent = nullptr);
     SimplexClass(const LpStructure& otherStructure);
     SimplexClass();
